fix: Url Service CRUD

+++ app/Services/UrlService.php
@@ -6,27 +6,40 @@ use App\Models\Url;
 
 class UrlService
 {
-    public function create(array $data): Url
+    public function create(array $data, int $userId): Url
     {
-        return Url::create($data);
-    }
+        $url = Url::query()->make($data);
+        $url->user_id = $userId;
+        $url->save();
 
-    public function read(int $id): ?Url
-    {
-        return Url::findOrFail($id);
-    }
-
-    public function update(array $data, int $id, ): Url
-    {
-        $url = Url::findOrFail($id);
-        $url->update($data);
         return $url;
     }
 
-    public function delete(int $id): bool
+    public function read(int $id, int $userId): ?Url
     {
-        $url = Url::findOrFail($id);
-        $url->delete();
-        return true;
+        return Url::query()->where("user_id", $userId)->findOrFail($id);
+    }
+
+    public function update(array $data, int $id, int $userId): ?Url
+    {
+        $url = Url::query()->where("user_id", $userId)->findOrFail($id);
+
+        if ($url) {
+            $url->update($data);
+        }
+
+        return $url;
+    }
+
+    public function delete(int $id, int $userId): bool
+    {
+        $url = Url::query()->where("user_id", $userId)->findOrFail($id);
+
+        if ($url) {
+            $url->delete();
+            return true;
+        } else {
+            return false;
+        }
     }
 }
